feat: add reorganization of the structure
