+++ infrastructure/currencies/currency-beacon.go
