+++ tests/cron-jobs/cronjob_integration_test.go
@@ -7,10 +7,11 @@ import (
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/golang/mock/gomock"
 	cronjobs "github.com/skobelina/currency_converter/domains/cron-jobs"
-	"github.com/skobelina/currency_converter/domains/mails"
-	"github.com/skobelina/currency_converter/domains/rates"
-	"github.com/skobelina/currency_converter/domains/subscribers"
-	"github.com/skobelina/currency_converter/repo"
+	subscriber "github.com/skobelina/currency_converter/domains/subscribers"
+	"github.com/skobelina/currency_converter/infrastructure/repo"
+	"github.com/skobelina/currency_converter/mocks/mails"
+	"github.com/skobelina/currency_converter/mocks/rates"
+	"github.com/skobelina/currency_converter/mocks/subscribers"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
@@ -42,8 +43,8 @@ func TestNotificationExchangeRates_Success(t *testing.T) {
 	mockRate := 27.32
 	mockRateService.EXPECT().Get().Return(&mockRate, nil)
 
-	mockSubscribers := &subscribers.SearchSubscribeResponse{
-		Data: []subscribers.Subscriber{
+	mockSubscribers := &subscriber.SearchSubscribeResponse{
+		Data: []subscriber.Subscriber{
 			{Email: "test1@example.com"},
 			{Email: "test2@example.com"},
 		},
@@ -72,7 +73,7 @@ func TestNotificationExchangeRates_Error(t *testing.T) {
 	mockRateService := rates.NewMockRateServiceInterface(ctrl)
 	mockSubscriberService := subscribers.NewMockSubscriberServiceInterface(ctrl)
 	mockRateService.EXPECT().Get().Return(nil, errors.New("internal server error"))
-	mockSubscriberService.EXPECT().Search(gomock.Any()).Return(&subscribers.SearchSubscribeResponse{}, nil)
+	mockSubscriberService.EXPECT().Search(gomock.Any()).Return(&subscriber.SearchSubscribeResponse{}, nil)
 	cronJobService := cronjobs.NewService(db, mockMailService, mockRateService, mockSubscriberService)
 	err := cronJobService.NotificationExchangeRates()
 	assert.Error(t, err)
