+++ tests/subscribers/handler_test.go
@@ -8,6 +8,7 @@ import (
 	"testing"
 
 	"github.com/gorilla/mux"
+	"github.com/skobelina/currency_converter/domains/subscribers"
 	utils "github.com/skobelina/currency_converter/utils/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
@@ -17,7 +18,7 @@ type MockService struct {
 	mock.Mock
 }
 
-func (m *MockService) Create(request *SubscriberRequest) (*string, error) {
+func (m *MockService) Create(request *subscribers.SubscriberRequest) (*string, error) {
 	args := m.Called(request)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
@@ -25,12 +26,12 @@ func (m *MockService) Create(request *SubscriberRequest) (*string, error) {
 	return args.Get(0).(*string), args.Error(1)
 }
 
-func (m *MockService) Search(filter *SearchSubscribeRequest) (*SearchSubscribeResponse, error) {
+func (m *MockService) Search(filter *subscribers.SearchSubscribeRequest) (*subscribers.SearchSubscribeResponse, error) {
 	args := m.Called(filter)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
 	}
-	return args.Get(0).(*SearchSubscribeResponse), args.Error(1)
+	return args.Get(0).(*subscribers.SearchSubscribeResponse), args.Error(1)
 }
 
 func TestHandler_Create_Success(t *testing.T) {
@@ -38,11 +39,11 @@ func TestHandler_Create_Success(t *testing.T) {
 	mockResponse := "Subscription created"
 	mockService.On("Create", mock.Anything).Return(&mockResponse, nil)
 
-	handler := NewHandler(mockService)
+	handler := subscribers.NewHandler(mockService)
 	router := mux.NewRouter()
 	handler.Register(router)
 
-	requestBody, err := json.Marshal(SubscriberRequest{})
+	requestBody, err := json.Marshal(subscribers.SubscriberRequest{})
 	assert.NoError(t, err)
 
 	req, err := http.NewRequest("POST", "/api/subscribe", bytes.NewBuffer(requestBody))
@@ -63,11 +64,11 @@ func TestHandler_Create_Conflict(t *testing.T) {
 	mockService := new(MockService)
 	mockService.On("Create", mock.Anything).Return(nil, utils.NewIsConflictError("conflict error"))
 
-	handler := NewHandler(mockService)
+	handler := subscribers.NewHandler(mockService)
 	router := mux.NewRouter()
 	handler.Register(router)
 
-	requestBody, err := json.Marshal(SubscriberRequest{})
+	requestBody, err := json.Marshal(subscribers.SubscriberRequest{})
 	assert.NoError(t, err)
 
 	req, err := http.NewRequest("POST", "/api/subscribe", bytes.NewBuffer(requestBody))
@@ -85,10 +86,10 @@ func TestHandler_Create_Conflict(t *testing.T) {
 
 func TestHandler_Search_Success(t *testing.T) {
 	mockService := new(MockService)
-	mockResponse := &SearchSubscribeResponse{}
+	mockResponse := &subscribers.SearchSubscribeResponse{}
 	mockService.On("Search", mock.Anything).Return(mockResponse, nil)
 
-	handler := NewHandler(mockService)
+	handler := subscribers.NewHandler(mockService)
 	router := mux.NewRouter()
 	handler.Register(router)
 
@@ -110,7 +111,7 @@ func TestHandler_Search_BadRequest(t *testing.T) {
 	mockService := new(MockService)
 	mockService.On("Search", mock.Anything).Return(nil, utils.NewBadRequestError("bad request"))
 
-	handler := NewHandler(mockService)
+	handler := subscribers.NewHandler(mockService)
 	router := mux.NewRouter()
 	handler.Register(router)
